arch: fix printf.c and virtio_disk.c for RV32
printf.c:
- Change printint parameter from 'long long' to 'long'
- Avoids 64-bit division which requires libgcc in RV32

virtio_disk.c:
- Set VirtIO descriptor high addresses to 0 (RV32 has 32-bit addresses)

+++ os/kernel/printf.c
@@ -26,11 +26,11 @@ static struct {
 static char digits[] = "0123456789abcdef";
 
 static void
-printint(long long xx, int base, int sign)
+printint(long xx, int base, int sign)
 {
   char buf[20];
   int i;
-  unsigned long long x;
+  unsigned long x;
 
   if(sign && (sign = (xx < 0)))
     x = -xx;
